+++ src/components/admin/ListLanguages.vue
@@ -1,44 +1,44 @@
 <template lang="pug">
   .ListLanguages
 
-    .columns
-      .column
-        h1.title.
-          Languages
-        h2.subtitle.
-          Languages description
-
-    .columns
-      .column
-        .box
-          table.table.is-striped.is-hoverable.is-fullwidth()
-            thead
-              tr
-                th.
-                  Languages Header 1
-                th.
-                  Languages Header 2
-                th.
-                  Actions
-            tbody
-              tr
-                td.
-                  content 1
-                td.
-                  content 2
-                td
-                  button.button.is-info.is-small.is-outlined()
-                    span.
-                      Edit
-                  .button.is-danger.is-small.is-outlined()
-                    span.
-                      Delete
+    admin-table(
+      :itemPlural='itemPlural',
+      :itemSingular='itemSingular',
+      :routeAddName='routeAddName',
+      :routeEditName='routeEditName',
+      :storeGetter='storeGetter',
+      :storeRemove='storeRemove',
+      :subtitle='subtitle',
+      :tableCols='tableCols',
+      :title='title',
+      )
 </template>
 
 <script>
+import AdminTable from '@/components/admin/ui/AdminTable'
 
 export default {
   name: 'ListLanguages',
+  components: {
+    AdminTable,
+  },
+  data () {
+    return {
+      itemPlural: 'Languages',
+      itemSingular: 'Language',
+      routeAddName: 'LanguageAdd',
+      routeEditName: 'LanguageEdit',
+      storeGetter: 'languagesList',
+      storeRemove: 'languageRemove',
+      subtitle: 'Languages description',
+      tableCols: [
+        { label: 'Name', prop: 'name' },
+        { label: 'Native name', prop: 'nativeName' },
+        { label: 'Code ISO_639-1', prop: 'code' },
+      ],
+      title: 'Languages',
+    }
+  },
 }
 </script>